fix: number entities now return int instead of float
This prevents decimal separators (e.g., commas) from appearing in localized UIs.
All configuration number entities (retention_days, max_deletes, keep_minimum_files,
max_files_in_folder) now correctly return integer values.

+++ custom_components/retention_cleaner/number.py
@@ -135,10 +135,10 @@ class RetentionCleanerNumberEntity(
         )
 
     @property
-    def native_value(self) -> float:
+    def native_value(self) -> int:
         """Return the current value from coordinator."""
         value = getattr(self.coordinator, self.entity_description.key)
-        return float(value)
+        return int(value)
 
     async def async_set_native_value(self, value: float) -> None:
         """Update the configuration value."""
